Change URL format for food posts
Previously, the continent was included in the URL. This seemed to
suggest that there would live a page there as well. As this is not
the case, just get rid of it and have it directly be /food/country
People will know where countries are.
Except for Americans: https://www.youtube.com/watch?v=kRh1zXFKC_o